requirements.txt
face-recognition==1.3.0
Pillow==10.0.0
numpy>=1.21.0
gTTS==2.3.2
Flask==3.0.2
Werkzeug==3.0.1